more fixes

+++ backend/Element/solid/moveable/MovableSolid.h
@@ -5,7 +5,7 @@
 #include "../../EmptyCell.h"
 #include "../../gas/Gas.h"
 #include "../../liquid/Liquid.h"
-#include "../../../Grid.cpp"
+#include "../../../Grid.h"
 #include <cmath>
 #include <random>
 #include <functional>
@@ -60,8 +60,8 @@ public:
 
     bool tryRandomDiagonalMovement(int x, int y, Grid& grid, std::function<void(int, int, int, int)> move, int step) {
         std::vector<Velocity> diagonals = (step % 2 == 0)
-            ? std::vector<Velocity>{{-1, std::floor(vel.y)}, {1, std::floor(vel.y)}} // Even step: Left then Right
-            : std::vector<Velocity>{{1, std::floor(vel.y)}, {-1, std::floor(vel.y)}}; // Odd step: Right then Left
+            ? std::vector<Velocity>{{-1, static_cast<int>(std::floor(vel.y))}, {1, static_cast<int>(std::floor(vel.y))}} // Even step: Left then Right
+            : std::vector<Velocity>{{1, static_cast<int>(std::floor(vel.y))}, {-1, static_cast<int>(std::floor(vel.y))}}; // Odd step: Right then Left
 
         for (const auto& d : diagonals) {
             if (tryMove(x, y, x + d.x, y + d.y, grid, move)) {
